Split best-for detail helpers
